+++ code/time_type_switch4.cpp
@@ -115,7 +115,7 @@ struct VisitorFor4 : ShapeVisitor
     const shape_kind<N>& first;
     const shape_kind<M>& second;
     const shape_kind<K>& third;
-    int                  result; 
+    size_t               result;
 };
 
 //------------------------------------------------------------------------------
@@ -135,7 +135,7 @@ struct VisitorFor3 : ShapeVisitor
     const shape_kind<N>& first;
     const shape_kind<M>& second;
     const Shape&         fourth;
-    int                  result; 
+    size_t               result;
 };
 
 //------------------------------------------------------------------------------
@@ -155,7 +155,7 @@ struct VisitorFor2 : ShapeVisitor
     const shape_kind<N>& first;
     const Shape&         third;
     const Shape&         fourth;
-    int                  result; 
+    size_t               result;
 };
 
 //------------------------------------------------------------------------------
@@ -174,7 +174,7 @@ struct VisitorFor1 : ShapeVisitor
     const Shape& second;
     const Shape& third;
     const Shape& fourth;
-    int          result;
+    size_t       result;
 };
 
 //------------------------------------------------------------------------------
